carried out the activity described in Readme.md 09-03-2022

+++ src/main/java/com/moviebook/service/impl/UserServiceImpl.java
@@ -1,15 +1,20 @@
 package com.moviebook.service.impl;
 
 import com.moviebook.domain.Role;
+import com.moviebook.dto.UserDto;
+import com.moviebook.dto.mapper.UserMapper;
 import com.moviebook.entities.User;
 import com.moviebook.repository.UserRepository;
 import com.moviebook.service.UserService;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -18,8 +23,11 @@ public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
 
-    public UserServiceImpl(UserRepository userRepository) {
+    private final UserMapper userMapper;
+
+    public UserServiceImpl(UserRepository userRepository, UserMapper userMapper) {
         this.userRepository = userRepository;
+        this.userMapper = userMapper;
     }
 
     @Override
@@ -29,33 +37,43 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User save(User user) {
-        user.setActivated(false);
-        user.setRol(Role.USER);
-        return this.userRepository.save(user);
+    public ResponseEntity<UserDto> save(UserDto newUser) {
+        if( newUser.getEmail().isBlank() || this.userRepository.existsById(newUser.getEmail()) )
+               return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+
+        newUser.setActivated(false);
+        newUser.setRol(Role.USER);
+        var userSaved = this.userRepository.save(this.userMapper.mapperDtoToFilmEntity(newUser));
+
+        return ResponseEntity.ok(this.userMapper.mapperEntityToUserDto(userSaved));
     }
 
     @Override
-    public User update(String userId, User user) {
+    public ResponseEntity<UserDto> update(String userId, UserDto userToUpdate) {
+        if (!this.userRepository.existsById(userId))
+               return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+
         var lastUser = this.userRepository.getById(userId);
-        user.setEmail(userId);
-        user.setRegistrationInformation(lastUser.getRegistration());
-        return this.userRepository.save(user);
+        userToUpdate.setEmail(userId);
+        userToUpdate.setRegistrationInformation(lastUser.getRegistration());
+        var userSaved = this.userRepository.save(this.userMapper.mapperDtoToFilmEntity(userToUpdate));
+
+        return ResponseEntity.ok(this.userMapper.mapperEntityToUserDto(userSaved));
     }
 
     @Override
-    public List<User> saveAll(List<User> users) {
-        return this.userRepository.saveAll(users);
-    }
+    public ResponseEntity<UserDto> deleteById(String id) {
+        if (!this.userRepository.existsById(id))
+            return  ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
 
-    @Override
-    public void deleteById(String id) {
         this.userRepository.deleteById(id);
+
+        return ResponseEntity.noContent().build();
     }
 
     @Override
-    public List<User> findAll() {
-        var users= this.userRepository.findAll();
+    public List<UserDto> findAll() {
+        var users= this.userRepository.findAll().stream().map(userMapper::mapperEntityToUserDto).collect(Collectors.toList());
         return users;
     }
     @Override
